Refactors Settings into its own module
